refactor: separate doc data and parsing
