+++ repo_agent/doc_meta_info.py
@@ -6,11 +6,9 @@ import json
 import os
 import threading
 from dataclasses import dataclass, field
-from enum import Enum, auto, unique
 from pathlib import Path
 from typing import Any, Callable, Dict, List, Optional
 
-import jedi
 from colorama import Fore, Style
 from prettytable import PrettyTable
 from tqdm import tqdm
@@ -20,275 +18,11 @@ from repo_agent.log import logger
 from repo_agent.multi_task_dispatch import Task, TaskManager
 from repo_agent.settings import SettingsManager
 from repo_agent.utils.meta_info_utils import latest_version_substring
+from repo_agent.doc_items import DocItem, DocItemStatus, DocItemType, need_to_generate
+from repo_agent.tree_builder import from_project_hierarchy_json
+from repo_agent.reference_resolver import parse_reference
 
 
-@unique
-class EdgeType(Enum):
-    reference_edge = auto()  # 一个obj引用另一个obj
-    subfile_edge = auto()    # 一个 文件/文件夹 属于一个文件夹
-    file_item_edge = auto()  # 一个 obj 属于一个文件
-
-
-@unique
-class DocItemType(Enum):
-    """
-    对可能的对象文档类型进行定义（分不同细粒度）
-    """
-    _repo = auto()          # 根节点，需要生成readme
-    _dir = auto()
-    _file = auto()
-    _class = auto()
-    _class_function = auto()
-    _function = auto()      # 文件内的常规function
-    _sub_function = auto()  # function内定义的subfunction
-    _global_var = auto()
-
-    def to_str(self):
-        # Convert enum member to a string representation used by the parser
-        if self == DocItemType._class:
-            return "ClassDef"
-        elif self in [DocItemType._function, DocItemType._class_function, DocItemType._sub_function]:
-            return "FunctionDef"
-        return self.name
-
-    def print_self(self):
-        # Return a color-coded name for pretty printing
-        color = Fore.WHITE
-        if self == DocItemType._dir:
-            color = Fore.GREEN
-        elif self == DocItemType._file:
-            color = Fore.YELLOW
-        elif self == DocItemType._class:
-            color = Fore.RED
-        elif self in [
-            DocItemType._function,
-            DocItemType._sub_function,
-            DocItemType._class_function,
-        ]:
-            color = Fore.BLUE
-        return color + self.name + Style.RESET_ALL
-
-
-@unique
-class DocItemStatus(Enum):
-    doc_up_to_date = auto()               # 无需生成文档
-    doc_has_not_been_generated = auto()   # 文档还未生成，需要生成
-    code_changed = auto()                 # 源码被修改，需要更新文档
-    add_new_referencer = auto()           # 添加了新的引用者
-    referencer_not_exist = auto()         # 曾经引用他的对象被删除或不再引用他了
-
-
-def need_to_generate(doc_item: DocItem, ignore_list: List[str] = []) -> bool:
-    """
-    只生成item的文档，文件及更高粒度都跳过。
-    如果属于一个黑名单文件也跳过。
-    """
-    if doc_item.item_status == DocItemStatus.doc_up_to_date:
-        return False
-
-    rel_file_path = doc_item.get_full_name()
-
-    # Skip doc generation for file/dir/repo
-    if doc_item.item_type in (DocItemType._file, DocItemType._dir, DocItemType._repo):
-        return False
-
-    parent = doc_item.father
-    while parent:
-        if parent.item_type == DocItemType._file:
-            # If the file path is in the ignore list, skip
-            if any(rel_file_path.startswith(ignore_item) for ignore_item in ignore_list):
-                return False
-            return True
-        parent = parent.father
-
-    return False
-
-
-@dataclass(eq=False)
-class DocItem:
-    item_type: DocItemType = DocItemType._class_function
-    item_status: DocItemStatus = DocItemStatus.doc_has_not_been_generated
-
-    obj_name: str = ""
-    code_start_line: int = -1
-    code_end_line: int = -1
-    md_content: List[str] = field(default_factory=list)         # 存储不同版本的doc
-    content: Dict[Any, Any] = field(default_factory=dict)       # 原本存储的信息
-
-    children: Dict[str, 'DocItem'] = field(default_factory=dict, compare=False)
-    father: Optional['DocItem'] = field(default=None, compare=False)
-
-    depth: int = 0
-    tree_path: List['DocItem'] = field(default_factory=list, compare=False)
-    max_reference_ansce: Optional['DocItem'] = field(default=None, compare=False)
-
-    reference_who: List['DocItem'] = field(default_factory=list, compare=False)       # 他引用了谁
-    who_reference_me: List['DocItem'] = field(default_factory=list, compare=False)    # 谁引用了他
-    special_reference_type: List[bool] = field(default_factory=list)
-
-    reference_who_name_list: List[str] = field(default_factory=list)
-    who_reference_me_name_list: List[str] = field(default_factory=list)
-
-    has_task: bool = False
-    multithread_task_id: int = -1
-
-    @staticmethod
-    def has_ans_relation(a: DocItem, b: DocItem):
-        """
-        Check if there is an ancestor relationship between two nodes.
-        Return whichever node is higher in the ancestor chain if found.
-        Otherwise return None.
-        """
-        if b in a.tree_path:
-            return b
-        if a in b.tree_path:
-            return a
-        return None
-
-    def get_travel_list(self):
-        """
-        Return a list of self + all descendants (pre-order traversal).
-        """
-        nodes = [self]
-        for child in self.children.values():
-            nodes.extend(child.get_travel_list())
-        return nodes
-
-    def check_depth(self):
-        """
-        Recursively calculate the depth of the node.
-        """
-        if len(self.children) == 0:
-            self.depth = 0
-            return 0
-        max_child_depth = 0
-        for child in self.children.values():
-            child_depth = child.check_depth()
-            max_child_depth = max(child_depth, max_child_depth)
-        self.depth = max_child_depth + 1
-        return self.depth
-
-    def parse_tree_path(self, now_path):
-        """
-        Recursively build the path from the root to this node.
-        """
-        self.tree_path = now_path + [self]
-        for child in self.children.values():
-            child.parse_tree_path(self.tree_path)
-
-    def get_file_name(self):
-        full_name = self.get_full_name()
-        return full_name.split(".py")[0] + ".py"
-
-    def get_full_name(self, strict=False):
-        """
-        获取从下到上所有的obj名字, 以 / 分隔.
-        If `strict=True`, handle potential naming conflicts by appending `(name_duplicate_version)`.
-        """
-        if self.father is None:
-            return self.obj_name
-
-        name_list = []
-        current = self
-        while current is not None:
-            current_name = current.obj_name
-            if strict and current.father is not None:
-                for name, item in current.father.children.items():
-                    if item == current:
-                        current_name = name
-                        break
-                if current_name != current.obj_name:
-                    current_name += "(name_duplicate_version)"
-            name_list.insert(0, current_name)
-            current = current.father
-
-        # The first node in the chain is the repo root; skip that name
-        name_list = name_list[1:]
-        return "/".join(name_list)
-
-    def find(self, recursive_file_path: list) -> Optional[DocItem]:
-        """
-        From the repo root, follow the path to find a nested child file or object.
-        Return None if not found.
-        """
-        assert self.item_type == DocItemType._repo
-        pos = 0
-        current_item = self
-        while pos < len(recursive_file_path):
-            if recursive_file_path[pos] not in current_item.children:
-                return None
-            current_item = current_item.children[recursive_file_path[pos]]
-            pos += 1
-        return current_item
-
-    @staticmethod
-    def check_has_task(current_item: DocItem, ignore_list: List[str] = []):
-        """
-        Mark doc items that need generation or whose children need generation.
-        """
-        if need_to_generate(current_item, ignore_list=ignore_list):
-            current_item.has_task = True
-        for child in current_item.children.values():
-            DocItem.check_has_task(child, ignore_list)
-            current_item.has_task = current_item.has_task or child.has_task
-
-    def print_recursive(self, indent=0, print_content=False, diff_status=False, ignore_list: List[str] = []):
-        """
-        递归打印repo对象信息
-        """
-        def print_indent(ind=0):
-            if ind == 0:
-                return ""
-            return "  " * ind + "|-"
-
-        item_name = self.obj_name
-        setting = SettingsManager.get_setting()
-
-        # If this is the repo root, display the actual folder name
-        if self.item_type == DocItemType._repo:
-            item_name = setting.project.target_repo
-
-        if diff_status and need_to_generate(self, ignore_list=ignore_list):
-            print(f"{print_indent(indent)}{self.item_type.print_self()}: {item_name} : {self.item_status.name}")
-        else:
-            print(f"{print_indent(indent)}{self.item_type.print_self()}: {item_name}")
-
-        for child_name, child in self.children.items():
-            # If diff_status is True, skip children that have no tasks
-            if diff_status and not child.has_task:
-                continue
-            child.print_recursive(
-                indent=indent + 1,
-                print_content=print_content,
-                diff_status=diff_status,
-                ignore_list=ignore_list,
-            )
-
-
-def find_all_referencer(repo_path, variable_name, file_path, line_number, column_number, in_file_only=False):
-    """
-    使用 jedi 查找对某个变量的引用位置。
-    """
-    script = jedi.Script(path=os.path.join(repo_path, file_path))
-    try:
-        if in_file_only:
-            references = script.get_references(line=line_number, column=column_number, scope="file")
-        else:
-            references = script.get_references(line=line_number, column=column_number)
-
-        variable_references = [ref for ref in references if ref.name == variable_name]
-        return [
-            (os.path.relpath(ref.module_path, repo_path), ref.line, ref.column)
-            for ref in variable_references
-            if not (ref.line == line_number and ref.column == column_number)
-        ]
-    except Exception as e:
-        logger.error(f"Error occurred: {e}")
-        logger.error(
-            f"Parameters: variable_name={variable_name}, file_path={file_path}, "
-            f"line_number={line_number}, column_number={column_number}"
-        )
-        return []
 
 
 @dataclass
@@ -437,89 +171,8 @@ class MetaInfo:
         return current_item
 
     def parse_reference(self):
-        """
-        双向提取所有引用关系
-        """
-        file_nodes = self.get_all_files()
-
-        white_list_file_names = []
-        white_list_obj_names = []
-        if self.white_list is not None:
-            white_list_file_names = [cont["file_path"] for cont in self.white_list]
-            white_list_obj_names = [cont["id_text"] for cont in self.white_list]
-
-        for file_node in tqdm(file_nodes, desc="parsing bidirectional reference"):
-            rel_file_path = file_node.get_full_name()
-
-            if rel_file_path in self.jump_files:
-                # Skip files that are in jump_files
-                continue
-
-            if white_list_file_names and (file_node.get_file_name() not in white_list_file_names):
-                # Skip files not in the white list (if white_list_file_names is set)
-                continue
-
-            ref_count = 0
-
-            def walk_file(current_item: DocItem):
-                nonlocal ref_count
-                in_file_only = False
-
-                # If there's a white list of obj_names, only parse references within file if not matched
-                if white_list_obj_names and (current_item.obj_name not in white_list_obj_names):
-                    in_file_only = True
-
-                reference_list = find_all_referencer(
-                    repo_path=self.repo_path,
-                    variable_name=current_item.obj_name,
-                    file_path=rel_file_path,
-                    line_number=current_item.content["code_start_line"],
-                    column_number=current_item.content["name_column"],
-                    in_file_only=in_file_only,
-                )
-
-                for referencer_pos in reference_list:
-                    referencer_file_ral_path = referencer_pos[0]
-
-                    # Skip references from unstaged/fake or untracked files
-                    if referencer_file_ral_path in self.fake_file_reflection.values():
-                        print(f"{Fore.LIGHTBLUE_EX}[Reference From Unstaged Version, skip]{Style.RESET_ALL} {referencer_file_ral_path} -> {current_item.get_full_name()}")
-                        continue
-                    if referencer_file_ral_path in self.jump_files:
-                        print(f"{Fore.LIGHTBLUE_EX}[Reference From Untracked Version, skip]{Style.RESET_ALL} {referencer_file_ral_path} -> {current_item.get_full_name()}")
-                        continue
-
-                    target_file_hiera = referencer_file_ral_path.split("/")
-                    referencer_file_item = self.target_repo_hierarchical_tree.find(target_file_hiera)
-                    if referencer_file_item is None:
-                        print(f'{Fore.LIGHTRED_EX}Error: Find "{referencer_file_ral_path}"(not in target repo){Style.RESET_ALL} referenced {current_item.get_full_name()}')
-                        continue
-
-                    referencer_node = self.find_obj_with_lineno(referencer_file_item, referencer_pos[1])
-
-                    # Skip name-duplicate references
-                    if referencer_node.obj_name == current_item.obj_name:
-                        logger.info(f"Jedi find {current_item.get_full_name()} with name_duplicate_reference, skipped")
-                        continue
-
-                    # If no ancestor relationship, record the reference
-                    if DocItem.has_ans_relation(current_item, referencer_node) is None:
-                        if current_item not in referencer_node.reference_who:
-                            special_reference = (
-                                referencer_node.item_type
-                                in [DocItemType._function, DocItemType._sub_function, DocItemType._class_function]
-                            ) and (referencer_node.code_start_line == referencer_pos[1])
-                            referencer_node.special_reference_type.append(special_reference)
-                            referencer_node.reference_who.append(current_item)
-                            current_item.who_reference_me.append(referencer_node)
-                            ref_count += 1
-
-                for child in current_item.children.values():
-                    walk_file(child)
-
-            for child in file_node.children.values():
-                walk_file(child)
-
+        """Delegate to :func:`reference_resolver.parse_reference`."""
+        parse_reference(self)
     def get_task_manager(self, current_item: DocItem, task_available_func) -> TaskManager:
         """
         先写一个退化的版本，只考虑拓扑引用关系
@@ -744,129 +397,9 @@ class MetaInfo:
         return hierachy_json
 
     @staticmethod
-    def from_project_hierarchy_json(project_hierarchy_json) -> MetaInfo:
-        """
-        将 project_hierarchy.json 转换到我们的数据结构
-        """
-        setting = SettingsManager.get_setting()
-        target_meta_info = MetaInfo(
-            target_repo_hierarchical_tree=DocItem(
-                item_type=DocItemType._repo,
-                obj_name="full_repo",
-            )
-        )
-
-        for file_name, file_content in tqdm(project_hierarchy_json.items(), desc="parsing parent relationship"):
-            full_path = os.path.join(setting.project.target_repo, file_name)
-            if not os.path.exists(full_path):
-                logger.info(f"deleted content: {file_name}")
-                continue
-            if os.path.getsize(full_path) == 0:
-                logger.info(f"blank content: {file_name}")
-                continue
-
-            recursive_file_path = file_name.split("/")
-            now_structure = target_meta_info.target_repo_hierarchical_tree
-
-            # Build directory chain if needed
-            for pos in range(len(recursive_file_path) - 1):
-                if recursive_file_path[pos] not in now_structure.children:
-                    now_structure.children[recursive_file_path[pos]] = DocItem(
-                        item_type=DocItemType._dir,
-                        md_content="",
-                        obj_name=recursive_file_path[pos],
-                    )
-                    now_structure.children[recursive_file_path[pos]].father = now_structure
-                now_structure = now_structure.children[recursive_file_path[pos]]
-
-            if recursive_file_path[-1] not in now_structure.children:
-                now_structure.children[recursive_file_path[-1]] = DocItem(
-                    item_type=DocItemType._file,
-                    obj_name=recursive_file_path[-1],
-                )
-                now_structure.children[recursive_file_path[-1]].father = now_structure
-
-            file_item = target_meta_info.target_repo_hierarchical_tree.find(recursive_file_path)
-            assert file_item.item_type == DocItemType._file
-
-            obj_item_list: List[DocItem] = []
-            for value in file_content:
-                obj_doc_item = DocItem(
-                    obj_name=value["name"],
-                    content=value,
-                    md_content=value["md_content"],
-                    code_start_line=value["code_start_line"],
-                    code_end_line=value["code_end_line"],
-                )
-                if "item_status" in value:
-                    obj_doc_item.item_status = DocItemStatus[value["item_status"]]
-                if "reference_who" in value:
-                    obj_doc_item.reference_who_name_list = value["reference_who"]
-                if "special_reference_type" in value:
-                    obj_doc_item.special_reference_type = value["special_reference_type"]
-                if "who_reference_me" in value:
-                    obj_doc_item.who_reference_me_name_list = value["who_reference_me"]
-                obj_item_list.append(obj_doc_item)
-
-            # Build parent-child relationships (O(n^2) approach)
-            for item in obj_item_list:
-                potential_father: Optional[DocItem] = None
-
-                def code_contain(child_item: DocItem, parent_candidate: DocItem) -> bool:
-                    if parent_candidate.code_end_line == child_item.code_end_line \
-                       and parent_candidate.code_start_line == child_item.code_start_line:
-                        return False
-                    if parent_candidate.code_end_line < child_item.code_end_line \
-                       or parent_candidate.code_start_line > child_item.code_start_line:
-                        return False
-                    return True
-
-                for other_item in obj_item_list:
-                    if code_contain(item, other_item):
-                        if potential_father is None or (
-                            (other_item.code_end_line - other_item.code_start_line)
-                            < (potential_father.code_end_line - potential_father.code_start_line)
-                        ):
-                            potential_father = other_item
-
-                if potential_father is None:
-                    potential_father = file_item
-
-                item.father = potential_father
-
-                child_name = item.obj_name
-                if child_name in potential_father.children:
-                    # Rename duplicates
-                    name_id = 0
-                    while (child_name + f"_{name_id}") in potential_father.children:
-                        name_id += 1
-                    logger.warning(
-                        f"Name duplicate in {file_item.get_full_name()}: rename to {item.obj_name}->{child_name}_{name_id}"
-                    )
-                    child_name += f"_{name_id}"
-
-                potential_father.children[child_name] = item
-
-            # Convert "ClassDef"/"FunctionDef" to the correct DocItemType
-            def change_items(node: DocItem):
-                if node.item_type != DocItemType._file:
-                    if node.content["type"] == "ClassDef":
-                        node.item_type = DocItemType._class
-                    elif node.content["type"] == "FunctionDef":
-                        node.item_type = DocItemType._function
-                        if node.father.item_type == DocItemType._class:
-                            node.item_type = DocItemType._class_function
-                        elif node.father.item_type in [DocItemType._function, DocItemType._sub_function]:
-                            node.item_type = DocItemType._sub_function
-                for c in node.children.values():
-                    change_items(c)
-
-            change_items(file_item)
-
-        # Build tree paths and calculate depths
-        target_meta_info.target_repo_hierarchical_tree.parse_tree_path(now_path=[])
-        target_meta_info.target_repo_hierarchical_tree.check_depth()
-        return target_meta_info
+    def from_project_hierarchy_json(project_hierarchy_json) -> "MetaInfo":
+        """Create :class:`MetaInfo` from ``project_hierarchy.json``."""
+        return from_project_hierarchy_json(project_hierarchy_json)
 
 
 if __name__ == "__main__":
